+++ nuxt.config.js
@@ -1,4 +1,5 @@
-module.exports = {
+export default {
+  target: 'static',
   /*
    ** Headers of the page
    */
@@ -10,19 +11,19 @@ module.exports = {
       {
         hid: 'description',
         name: 'description',
-        content: process.env.npm_package_description || ''
-      }
+        content: process.env.npm_package_description || '',
+      },
     ],
     link: [
       {
         rel: 'icon',
         type: 'image/x-icon',
-        href: '/favicon.ico'
-      }
+        href: '/favicon.ico',
+      },
     ],
     bodyAttrs: {
-      'data-version': process.env.npm_package_version || ''
-    }
+      'data-version': process.env.npm_package_version || '',
+    },
   },
   /*
    ** Customize the progress-bar color
@@ -32,16 +33,28 @@ module.exports = {
    ** Global CSS
    */
   css: [
+    { src: '@/assets/css/tailwind.css', lang: 'css' },
     { src: '@/assets/sass/app.scss', lang: 'scss' },
-    { src: '~/node_modules/highlight.js/styles/a11y-dark.css', lang: 'css' }
+    { src: '~/node_modules/highlight.js/styles/a11y-dark.css', lang: 'css' },
   ],
   /*
    ** Plugins to load before mounting the App
    */
   plugins: [],
-  /*
-   ** Nuxt.js modules
-   */
+
+  // Auto import components: https://go.nuxtjs.dev/config-components
+  components: true,
+
+  // Modules for dev and build (recommended): https://go.nuxtjs.dev/config-modules
+  buildModules: [
+    // https://go.nuxtjs.dev/eslint
+    '@nuxtjs/eslint-module',
+    // https://go.nuxtjs.dev/tailwindcss
+    '@nuxtjs/tailwindcss',
+    '@nuxtjs/markdownit',
+  ],
+
+  // Modules: https://go.nuxtjs.dev/config-modules
   modules: [
     // Doc:https://github.com/nuxt-community/modules/tree/master/packages/bulma
     // '@nuxtjs/bulma',
@@ -51,11 +64,11 @@ module.exports = {
     [
       '@nuxtjs/google-analytics',
       {
-        id: 'UA-26457988-4'
-      }
+        id: 'UA-26457988-4',
+      },
     ],
-    '@nuxtjs/markdownit',
-    '@nuxtjs/pwa'
+
+    '@nuxtjs/pwa',
   ],
   /*
    ** Axios module configuration
@@ -67,10 +80,15 @@ module.exports = {
     linkify: false,
     breaks: false,
     // use: [['markdown-it-container', 'name'], 'markdown-it-attrs']
-    use: ['markdown-it-highlightjs', 'markdown-it-attrs']
+    use: ['markdown-it-highlightjs', 'markdown-it-attrs'],
+  },
+  pwa: {
+    manifest: {
+      lang: 'en',
+    },
   },
   env: {
-    app_version: process.env.npm_package_version
+    app_version: process.env.npm_package_version,
   },
   /*
    ** Build configuration
@@ -79,13 +97,13 @@ module.exports = {
     postcss: {
       preset: {
         features: {
-          customProperties: false
-        }
-      }
+          customProperties: false,
+        },
+      },
     },
     /*
      ** You can extend webpack config here
      */
-    extend(config, ctx) {}
-  }
+    extend(config, ctx) {},
+  },
 }