+++ pages/index.vue
@@ -1,14 +1,16 @@
 <template lang="pug">
   div.home
-    section.hero.is-medium.is-bold.is-dark
-      .hero-body
-        .container
-          h1.title.is-1(v-html="pageTitle")
-          h2.subtitle(v-html="subtitle")
+    section.hero.bg-gray-500.mb-16
+      .hero-body.max-w-7xl
+        .container.mx-auto
+          .py-16
+            h1.title.text-gray-100 {{ pageTitle }}
+            h2.subtitle.text-gray-100 {{ subtitle }}
     section.section
-      .container
-        article#intro.content.is-medium(v-html="intro")
-        article#toc.content.is-medium(v-html="toc")
+      .container.mx-auto
+        article
+          #intro.content.is-medium(v-html="intro")
+          #toc.content.is-medium(v-html="toc")
 </template>
 
 <script>
@@ -25,6 +27,11 @@ export default {
       title: [this.title, this.subtitle].join(' - '),
     }
   },
+  head() {
+    return {
+      title: this.title,
+    }
+  },
   computed: {
     intro() {
       return intro
@@ -36,11 +43,6 @@ export default {
   async mounted() {
     this.title = (await this.pageTitle) + ' - ' + this.subtitle
   },
-  head() {
-    return {
-      title: this.title,
-    }
-  },
 }
 </script>
 
@@ -48,5 +50,13 @@ export default {
 .section > .container {
   background-color: white;
   padding: 0 1em;
+}
+
+.title {
+  @apply text-4xl sm:text-5xl;
+}
+
+.subtitle {
+  @apply text-2xl sm:text-4xl;
 }
 </style>